requirements.txt
Django==5.1.3
gunicorn==23.0.0
whitenoise==6.8.2
django-debug-toolbar==4.4.6

psycopg==3.2.3

redis==5.2.0
celery==5.4.0

flake8==7.1.1
isort==5.13.2
black==24.10.0

setuptools==74.1.2

djangorestframework==3.15.2

requests==2.26.0
matplotlib==3.10.0
pandas==2.2.3